+++ src/render/widget.rs
@@ -19,6 +19,7 @@ use sdl2::video::Window;
 
 use crate::render::callbacks::*;
 use crate::render::layout_cache::LayoutContainer;
+use crate::render::texture_cache::TextureCache;
 use crate::render::texture_store::TextureStore;
 use crate::render::widget_cache::WidgetContainer;
 use crate::render::widget_config::*;
@@ -27,7 +28,6 @@ use sdl2::event::Event;
 use sdl2::pixels::Color;
 use std::any::Any;
 use std::collections::HashMap;
-use crate::render::texture_cache::TextureCache;
 
 /// This trait is shared by all `Widget` objects that have a presence on the screen.  Functions that
 /// must be implemented are documented in the trait.