+++ src/widgets/checkbox_widget.rs
@@ -26,6 +26,7 @@ use sdl2::video::Window;
 
 use crate::render::canvas_helper::CanvasHelper;
 use crate::render::layout_cache::LayoutContainer;
+use crate::render::texture_cache::TextureCache;
 use crate::render::texture_store::TextureStore;
 use crate::render::widget_config::CompassPosition::Center;
 use crate::widgets::image_widget::ImageWidget;
@@ -33,7 +34,6 @@ use crate::widgets::text_widget::{TextJustify, TextWidget};
 use sdl2::pixels::Color;
 use std::any::Any;
 use std::collections::HashMap;
-use crate::render::texture_cache::TextureCache;
 
 /// This is the callback type that is used when an `on_toggle` callback is triggered from this
 /// `Widget`.