+++ src/widgets/grid_widget.rs
@@ -24,12 +24,12 @@ use sdl2::video::Window;
 
 use crate::render::canvas_helper::CanvasHelper;
 use crate::render::layout_cache::LayoutContainer;
+use crate::render::texture_cache::TextureCache;
 use crate::render::texture_store::TextureStore;
 use sdl2::pixels::Color;
 use sdl2::rect::{Point, Rect};
 use std::any::Any;
 use std::collections::HashMap;
-use crate::render::texture_cache::TextureCache;
 
 /// This is the storage object for the `GridWidget`.  It stores the config, properties, callback registry.
 pub struct GridWidget {
@@ -94,24 +94,20 @@ impl Widget for GridWidget {
                     texture.set_draw_color(Color::RGB(192, 192, 192));
 
                     for i in (0..size[SIZE_WIDTH]).step_by(grid_size) {
-                        texture.draw_line(
-                            Point::new(i as i32, 0),
-                            Point::new(
-                                i as i32,
-                                size[SIZE_HEIGHT] as i32
-                            ),
-                        )
+                        texture
+                            .draw_line(
+                                Point::new(i as i32, 0),
+                                Point::new(i as i32, size[SIZE_HEIGHT] as i32),
+                            )
                             .unwrap();
                     }
 
                     for i in (0..size[SIZE_HEIGHT]).step_by(grid_size) {
-                        texture.draw_line(
-                            Point::new(0, i as i32),
-                            Point::new(
-                                size[SIZE_WIDTH] as i32,
-                                i as i32,
-                            ),
-                        )
+                        texture
+                            .draw_line(
+                                Point::new(0, i as i32),
+                                Point::new(size[SIZE_WIDTH] as i32, i as i32),
+                            )
                             .unwrap();
                     }
                 } else {
@@ -125,14 +121,11 @@ impl Widget for GridWidget {
                 }
 
                 texture.set_draw_color(border_color);
-                texture.draw_rect(Rect::new(
-                    0,
-                    0,
-                    size[0],
-                    size[1],
-                ))
+                texture
+                    .draw_rect(Rect::new(0, 0, size[0], size[1]))
                     .unwrap();
-            }).unwrap();
+            })
+            .unwrap();
         }
 
         self.texture_store.get_optional_ref()