+++ src/widgets/image_widget.rs
@@ -27,11 +27,11 @@ use sdl2::rect::Rect;
 use sdl2::render::{Canvas, Texture, TextureQuery};
 use sdl2::video::Window;
 
+use crate::render::texture_cache::TextureCache;
 use crate::render::texture_store::TextureStore;
 use std::any::Any;
 use std::collections::HashMap;
 use std::path::Path;
-use crate::render::texture_cache::TextureCache;
 
 /// This is the storage object for the `ImageWidget`.  It stores the config, properties, callback registry,
 /// the image name, and a scale flag.
@@ -70,16 +70,13 @@ impl ImageWidget {
 /// This is the `Widget` implementation of the `ImageWidget`.  Image is rendered onto a 3D texture, then
 /// copied to the canvas after rendering.
 impl Widget for ImageWidget {
-    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
+    fn draw(&mut self, c: &mut Canvas<Window>, t: &mut TextureCache) -> Option<&Texture> {
         let base_color = self.get_color(CONFIG_COLOR_BASE);
 
         c.set_draw_color(base_color);
         c.fill_rect(self.get_drawing_area()).unwrap();
 
-        let texture_creator = c.texture_creator();
-        let texture = texture_creator
-            .load_texture(Path::new(&self.image_name))
-            .unwrap();
+        let texture = t.get_image(c, self.image_name.clone());
         let widget_w = self.get_size(CONFIG_SIZE)[0] as i32;
         let widget_h = self.get_size(CONFIG_SIZE)[1] as i32;
         let TextureQuery { width, height, .. } = texture.query();