+++ src/widgets/list_widget.rs
@@ -23,6 +23,7 @@ use sdl2::video::Window;
 
 use crate::render::canvas_helper::CanvasHelper;
 use crate::render::layout_cache::LayoutContainer;
+use crate::render::texture_cache::TextureCache;
 use crate::render::texture_store::TextureStore;
 use crate::render::{Points, Size, POINT_Y};
 use sdl2::pixels::Color;
@@ -30,7 +31,6 @@ use sdl2::rect::Rect;
 use std::any::Any;
 use std::collections::HashMap;
 use std::path::Path;
-use crate::render::texture_cache::TextureCache;
 
 /// This is the callback type that is used when an `on_selected` callback is triggered from this
 /// `Widget`.